+++ src/System/Taffybar/Widget.hs
@@ -75,6 +75,7 @@ module System.Taffybar.Widget
 
   -- * "System.Taffybar.Widget.SNITray"
   , module System.Taffybar.Widget.SNITray
+  , module System.Taffybar.Widget.SNITray.PrioritizedCollapsible
 
   -- * "System.Taffybar.Widget.SimpleClock"
   , module System.Taffybar.Widget.SimpleClock
@@ -161,6 +162,7 @@ import System.Taffybar.Widget.NetworkGraph
 import System.Taffybar.Widget.NetworkManager
 import System.Taffybar.Widget.PowerProfiles
 import System.Taffybar.Widget.SNITray
+import System.Taffybar.Widget.SNITray.PrioritizedCollapsible
 import System.Taffybar.Widget.SimpleClock
 import System.Taffybar.Widget.SimpleCommandButton
 import System.Taffybar.Widget.WakeupDebug